Move to Girder 3.
This removes the options to ingest from tcga-data.nci.nih.gov, as that
has not been served in a compatible manner in a long time.

Still needed:

- Add circle-ci testing

- Add deployment to pypi on successful builds

- Migrate non-toolkit code from HistomicsTK

- Migrate deployment from HistomicsTK

- Migrate docs from HistomicsTK
