+++ .editorconfig
@@ -1,0 +1,27 @@
+root = true
+
+[*]
+indent_style = space
+end_of_line = lf
+insert_final_newline = true
+trim_trailing_whitespace = true
+charset = utf-8
+
+[*.py]
+indent_size = 4
+max_line_length = 100
+
+[*.js]
+indent_size = 4
+
+[*.json]
+indent_size = 4
+
+[*.pug]
+indent_size = 2
+
+[*.styl]
+indent_size = 2
+
+[*.yml]
+indent_size = 2
