+++ digital_slide_archive/models/pathology.py
