+++ digital_slide_archive/web_client/main.js
@@ -1,0 +1,1 @@
+console.log('Loaded Digital Slide Archive!');
