+++ digital_slide_archive/web_client/package.json
@@ -1,0 +1,154 @@
+{
+    "name": "@girder/digital_slide_archive",
+    "version": "1.0.0",
+    "private": true,
+    "description": "Organize, visualize, and analyze histology images.",
+    "homepage": "https://github.com/DigitalSlideArchive/digital_slide_archive",
+    "license": "Apache-2.0",
+    "peerDependencies": {
+        "@girder/core": "*"
+    },
+    "girderPlugin": {
+        "name": "digital_slide_archive",
+        "main": "./main.js"
+    },
+    "scripts": {
+        "lint": "eslint . && pug-lint . && stylint"
+    },
+    "devDependencies": {
+        "@girder/eslint-config": "*",
+        "@girder/pug-lint-config": "*",
+        "eslint": "^5",
+        "eslint-config-semistandard": "^13",
+        "eslint-config-standard": "^12",
+        "eslint-plugin-backbone": "*",
+        "eslint-plugin-import": "*",
+        "eslint-plugin-node": "*",
+        "eslint-plugin-promise": "*",
+        "eslint-plugin-standard": "*",
+        "eslint-plugin-underscore": "*",
+        "pug-lint": "^2",
+        "stylint": "^2"
+    },
+    "eslintConfig": {
+        "extends": "@girder",
+        "root": true
+    },
+    "pugLintConfig": {
+        "extends": "@girder/pug-lint-config"
+    },
+    "stylintrc": {
+        "blocks": false,
+        "brackets": {
+            "expect": "never",
+            "error": true
+        },
+        "colons": {
+            "expect": "never",
+            "error": true
+        },
+        "colors": false,
+        "commaSpace": {
+            "expect": "always",
+            "error": true
+        },
+        "commentSpace": {
+            "expect": "always",
+            "error": true
+        },
+        "cssLiteral": {
+            "expect": "never",
+            "error": true
+        },
+        "depthLimit": false,
+        "duplicates": {
+            "expect": true,
+            "error": true
+        },
+        "efficient": {
+            "expect": "always",
+            "error": true
+        },
+        "exclude": [
+            "node_modules/**"
+        ],
+        "extendPref": "@extend",
+        "globalDupe": false,
+        "groupOutputByFile": {
+            "expect": true,
+            "error": true
+        },
+        "indentPref": {
+            "expect": 2,
+            "error": true
+        },
+        "leadingZero": {
+            "expect": "always",
+            "error": true
+        },
+        "maxErrors": false,
+        "maxWarnings": false,
+        "mixed": false,
+        "mixins": [],
+        "namingConvention": false,
+        "namingConventionStrict": false,
+        "none": {
+            "expect": "always",
+            "error": true
+        },
+        "noImportant": false,
+        "parenSpace": {
+            "expect": "never",
+            "error": true
+        },
+        "placeholders": false,
+        "prefixVarsWithDollar": {
+            "expect": "always",
+            "error": true
+        },
+        "quotePref": {
+            "expect": "double",
+            "error": true
+        },
+        "reporterOptions": {
+            "columns": [
+                "lineData",
+                "severity",
+                "description",
+                "rule"
+            ],
+            "columnSplitter": "  ",
+            "showHeaders": false,
+            "truncate": true
+        },
+        "semicolons": {
+            "expect": "never",
+            "error": true
+        },
+        "sortOrder": false,
+        "stackedProperties": {
+            "expect": "never",
+            "error": true
+        },
+        "trailingWhitespace": {
+            "expect": "never",
+            "error": true
+        },
+        "universal": {
+            "expect": "never",
+            "error": true
+        },
+        "valid": {
+            "expect": true,
+            "error": true
+        },
+        "zeroUnits": {
+            "expect": "never",
+            "error": true
+        },
+        "zIndexNormalize": {
+            "expect": 5,
+            "error": true
+        }
+    }
+}
