--- requirements.txt
@@ -1,1 +1,0 @@
-lxml>=3.4.4
