--- server/datasets/tcga/constants.py
@@ -1,950 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-###############################################################################
-#  Copyright Kitware Inc.
-#
-#  Licensed under the Apache License, Version 2.0 ( the "License" );
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-###############################################################################
-
-# flake8: noqa: E501
-
-
-class TcgaCodes(object):
-
-    DISEASE_STUDIES = {
-        # 'Study Abbreviation': 'Study Name',
-        'LAML': 'Acute Myeloid Leukemia',
-        'ACC': 'Adrenocortical carcinoma',
-        'BLCA': 'Bladder Urothelial Carcinoma',
-        'LGG': 'Brain Lower Grade Glioma',
-        'BRCA': 'Breast invasive carcinoma',
-        'CESC': 'Cervical squamous cell carcinoma and endocervical adenocarcinoma',
-        'CHOL': 'Cholangiocarcinoma',
-        'LCML': 'Chronic Myelogenous Leukemia',
-        'COAD': 'Colon adenocarcinoma',
-        'CNTL': 'Controls',
-        'ESCA': 'Esophageal carcinoma ',
-        'FPPP': 'FFPE Pilot Phase II',
-        'GBM': 'Glioblastoma multiforme',
-        'HNSC': 'Head and Neck squamous cell carcinoma',
-        'KICH': 'Kidney Chromophobe',
-        'KIRC': 'Kidney renal clear cell carcinoma',
-        'KIRP': 'Kidney renal papillary cell carcinoma',
-        'LIHC': 'Liver hepatocellular carcinoma',
-        'LUAD': 'Lung adenocarcinoma',
-        'LUSC': 'Lung squamous cell carcinoma',
-        'DLBC': 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma',
-        'MESO': 'Mesothelioma',
-        'MISC': 'Miscellaneous',
-        'OV': 'Ovarian serous cystadenocarcinoma',
-        'PAAD': 'Pancreatic adenocarcinoma',
-        'PCPG': 'Pheochromocytoma and Paraganglioma',
-        'PRAD': 'Prostate adenocarcinoma',
-        'READ': 'Rectum adenocarcinoma',
-        'SARC': 'Sarcoma',
-        'SKCM': 'Skin Cutaneous Melanoma',
-        'STAD': 'Stomach adenocarcinoma',
-        'TGCT': 'Testicular Germ Cell Tumors',
-        'THYM': 'Thymoma',
-        'THCA': 'Thyroid carcinoma',
-        'UCS': 'Uterine Carcinosarcoma',
-        'UCEC': 'Uterine Corpus Endometrial Carcinoma',
-        'UVM': 'Uveal Melanoma',
-    }
-
-    REPOSITORY_TYPES = {
-        'bcr',  # 'Biospecimen Core Resource'
-        'cgcc',
-        'gsc',
-    }
-
-    DATA_PROVIDERS = {
-        'biotab',  # Clinical metadata, skip
-        'intgen.org',
-        'nationwidechildrens.org',
-        'genome.wustl.edu',
-        'supplemental'  # unknown, appears under 'tumor/ov/bcr/', skip
-    }
-
-    DATA_TYPES = {
-        'bio',  # XML format clinical metadata, skip
-        'biotab',  # CSV format clinical metadata, skip
-        'pathology_reports',  # PDF format pathology reports, skip
-        'diagnostic_images',  # SVS format images, use
-        'tissue_images',  # SVS format images, use
-        'minbio'  # unknown, appears under 'tumor/gbm/bcr/intgen.org/', skip
-    }
-
-    SLIDE_LOCATION = {
-        'TS': 'Top Slide',
-        'MS': 'Middle Slide',
-        'BS': 'Bottom Slide',
-        'DX': 'Top Slide',
-    }
-
-    TISSUE_SOURCE_SITE = {
-        # 'TSS Code': ('Source Site', 'Study Name', 'BCR'),
-        '01': ('International Genomics Consortium', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '02': ('MD Anderson Cancer Center', 'Glioblastoma multiforme', 'IGC'),
-        '04': ('Gynecologic Oncology Group', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '05': ('Indivumed', 'Lung adenocarcinoma', 'IGC'),
-        '06': ('Henry Ford Hospital', 'Glioblastoma multiforme', 'IGC'),
-        '07': ('TGen', 'Cell Line Control', 'IGC'),
-        '08': ('UCSF', 'Glioblastoma multiforme', 'IGC'),
-        '09': ('UCSF', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '10': ('MD Anderson Cancer Center', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '11': ('MD Anderson Cancer Center', 'Lung squamous cell carcinoma', 'IGC'),
-        '12': ('Duke', 'Glioblastoma multiforme', 'IGC'),
-        '13': ('Memorial Sloan Kettering', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '14': ('Emory University', 'Glioblastoma multiforme', 'IGC'),
-        '15': ('Mayo Clinic - Rochester', 'Glioblastoma multiforme', 'IGC'),
-        '16': ('Toronto Western Hospital', 'Glioblastoma multiforme', 'IGC'),
-        '17': ('Washington University', 'Lung adenocarcinoma', 'IGC'),
-        '18': ('Princess Margaret Hospital (Canada)', 'Lung squamous cell carcinoma', 'IGC'),
-        '19': ('Case Western', 'Glioblastoma multiforme', 'IGC'),
-        '1Z': ('Johns Hopkins', 'Thymoma', 'NCH'),
-        '20': ('Fox Chase Cancer Center', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '21': ('Fox Chase Cancer Center', 'Lung squamous cell carcinoma', 'IGC'),
-        '22': ('Mayo Clinic - Rochester', 'Lung squamous cell carcinoma', 'IGC'),
-        '23': ('Cedars Sinai', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '24': ('Washington University', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '25': ('Mayo Clinic - Rochester', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '26': ('University of Florida', 'Glioblastoma multiforme', 'IGC'),
-        '27': ('Milan - Italy, Fondazione IRCCS Instituto Neuroligico C. Besta', 'Glioblastoma multiforme', 'IGC'),
-        '28': ('Cedars Sinai', 'Glioblastoma multiforme', 'IGC'),
-        '29': ('Duke', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '2A': ('Memorial Sloan Kettering Cancer Center', 'Prostate adenocarcinoma', 'NCH'),
-        '2E': ('University of Kansas Medical Center', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        '2F': ('Erasmus MC', 'Bladder Urothelial Carcinoma', 'NCH'),
-        '2G': ('Erasmus MC', 'Testicular Germ Cell Tumors', 'NCH'),
-        '2H': ('Erasmus MC', 'Esophageal carcinoma ', 'NCH'),
-        '2J': ('Mayo Clinic', 'Pancreatic adenocarcinoma', 'NCH'),
-        '2K': ('Greenville Health System', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        '2L': ('Technical University of Munich', 'Pancreatic adenocarcinoma', 'NCH'),
-        '2M': ('Technical University of Munich', 'Esophageal carcinoma ', 'NCH'),
-        '2N': ('Technical University of Munich', 'Stomach adenocarcinoma', 'NCH'),
-        '2P': ('University of California San Diego', 'Pancreatic adenocarcinoma', 'NCH'),
-        '2V': ('University of California San Diego', 'Liver hepatocellular carcinoma', 'NCH'),
-        '2W': ('University of New Mexico', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        '2X': ('ABS IUPUI', 'Testicular Germ Cell Tumors', 'NCH'),
-        '2Y': ('Moffitt Cancer Center', 'Liver hepatocellular carcinoma', 'NCH'),
-        '2Z': ('Moffitt Cancer Center', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        '30': ('Harvard', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '31': ('Imperial College', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '32': ('St. Joseph\'s Hospital (AZ)', 'Glioblastoma multiforme', 'IGC'),
-        '33': ('Johns Hopkins', 'Lung squamous cell carcinoma', 'IGC'),
-        '34': ('University of Pittsburgh', 'Lung squamous cell carcinoma', 'IGC'),
-        '35': ('Cureline', 'Lung adenocarcinoma', 'IGC'),
-        '36': ('BC Cancer Agency', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '37': ('Cureline', 'Lung squamous cell carcinoma', 'IGC'),
-        '38': ('UNC', 'Lung adenocarcinoma', 'IGC'),
-        '39': ('MSKCC', 'Lung squamous cell carcinoma', 'IGC'),
-        '3A': ('Moffitt Cancer Center', 'Pancreatic adenocarcinoma', 'NCH'),
-        '3B': ('Moffitt Cancer Center', 'Sarcoma', 'NCH'),
-        '3C': ('Columbia University', 'Breast invasive carcinoma', 'NCH'),
-        '3E': ('Columbia University', 'Pancreatic adenocarcinoma', 'NCH'),
-        '3G': ('MD Anderson Cancer Center', 'Thymoma', 'NCH'),
-        '3H': ('MD Anderson Cancer Center', 'Mesothelioma', 'NCH'),
-        '3J': ('Carle Cancer Center', 'Breast invasive carcinoma', 'NCH'),
-        '3K': ('Boston Medical Center', 'Liver hepatocellular carcinoma', 'NCH'),
-        '3L': ('Albert Einstein Medical Center', 'Colon adenocarcinoma', 'NCH'),
-        '3M': ('University of Kansas Medical Center', 'Stomach adenocarcinoma', 'NCH'),
-        '3N': ('Greenville Health System', 'Skin Cutaneous Melanoma', 'NCH'),
-        '3P': ('Greenville Health System', 'Ovarian serous cystadenocarcinoma', 'NCH'),
-        '3Q': ('Greenville Health Systems', 'Thymoma', 'NCH'),
-        '3R': ('University of New Mexico', 'Sarcoma', 'NCH'),
-        '3S': ('University of New Mexico', 'Thymoma', 'NCH'),
-        '3T': ('Emory University', 'Thymoma', 'NCH'),
-        '3U': ('University of Chicago', 'Mesothelioma', 'NCH'),
-        '3W': ('University of California San Diego', 'Sarcoma', 'NCH'),
-        '3X': ('Alberta Health Services', 'Cholangiocarcinoma', 'NCH'),
-        '3Z': ('Mary Bird Perkins Cancer Center - Our Lady of the Lake', 'Kidney renal clear cell carcinoma', 'NCH'),
-        '41': ('Christiana Healthcare', 'Glioblastoma multiforme', 'IGC'),
-        '42': ('Christiana Healthcare', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '43': ('Christiana Healthcare', 'Lung squamous cell carcinoma', 'IGC'),
-        '44': ('Christiana Healthcare', 'Lung adenocarcinoma', 'IGC'),
-        '46': ('St. Joseph\'s Medical Center (MD)', 'Lung squamous cell carcinoma', 'IGC'),
-        '49': ('Johns Hopkins', 'Lung adenocarcinoma', 'IGC'),
-        '4A': ('Mary Bird Perkins Cancer Center - Our Lady of the Lake', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        '4B': ('Mary Bird Perkins Cancer Center - Our Lady of the Lake', 'Lung adenocarcinoma', 'NCH'),
-        '4C': ('Mary Bird Perkins Cancer Center - Our Lady of the Lake', 'Thyroid carcinoma', 'NCH'),
-        '4D': ('Molecular Response', 'Ovarian serous cystadenocarcinoma', 'NCH'),
-        '4E': ('Molecular Response', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        '4G': ('Sapienza University of Rome', 'Cholangiocarcinoma', 'NCH'),
-        '4H': ('Proteogenex, Inc.', 'Breast invasive carcinoma', 'NCH'),
-        '4J': ('Proteogenex, Inc.', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        '4K': ('Proteogenex, Inc.', 'Testicular Germ Cell Tumors', 'NCH'),
-        '4L': ('Proteogenex, Inc.', 'Prostate adenocarcinoma', 'NCH'),
-        '4N': ('Mary Bird Perkins Cancer Center - Our Lady of the Lake', 'Colon adenocarcinoma', 'NCH'),
-        '4P': ('Duke University', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        '4Q': ('Duke University', 'Sarcoma', 'NCH'),
-        '4R': ('Duke University', 'Liver hepatocellular carcinoma', 'NCH'),
-        '4S': ('Duke University', 'Prostate adenocarcinoma', 'NCH'),
-        '4T': ('Duke University', 'Colon adenocarcinoma', 'NCH'),
-        '4V': ('Hospital Louis Pradel', 'Thymoma', 'NCH'),
-        '4W': ('University of Miami', 'Glioblastoma multiforme', 'NCH'),
-        '4X': ('Yale University', 'Thymoma', 'NCH'),
-        '4Y': ('Medical College of Wisconsin', 'Sarcoma', 'NCH'),
-        '4Z': ('Barretos Cancer Hospital', 'Bladder Urothelial Carcinoma', 'NCH'),
-        '50': ('University of Pittsburgh', 'Lung adenocarcinoma', 'IGC'),
-        '51': ('UNC', 'Lung squamous cell carcinoma', 'IGC'),
-        '52': ('University of Miami', 'Lung squamous cell carcinoma', 'IGC'),
-        '53': ('University of Miami', 'Lung adenocarcinoma', 'IGC'),
-        '55': ('International Genomics Consortium', 'Lung adenocarcinoma', 'IGC'),
-        '56': ('International Genomics Consortium', 'Lung squamous cell carcinoma', 'IGC'),
-        '57': ('International Genomics Consortium', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '58': ('Thoraxklinik at University Hospital Heidelberg', 'Lung squamous cell carcinoma', 'IGC'),
-        '59': ('Roswell Park', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '5A': ('Wake Forest University', 'Cholangiocarcinoma', 'NCH'),
-        '5B': ('Medical College of Wisconsin', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        '5C': ('Cureline', 'Liver hepatocellular carcinoma', 'NCH'),
-        '5D': ('University of Miami', 'Sarcoma', 'NCH'),
-        '5F': ('Duke University', 'Thyroid carcinoma', 'NCH'),
-        '5G': ('Cleveland Clinic Foundation', 'Thymoma', 'NCH'),
-        '5H': ('Retina Consultants Houston', 'Uveal Melanoma', 'NCH'),
-        '5J': ('Cureline', 'Acute Myeloid Leukemia', 'NCH'),
-        '5K': ('St. Joseph\'s Hospital AZ', 'Thymoma', 'NCH'),
-        '5L': ('University of Sao Paulo', 'Breast invasive carcinoma', 'NCH'),
-        '5M': ('University of Sao Paulo', 'Colon adenocarcinoma', 'NCH'),
-        '5N': ('University Hospital Erlangen', 'Bladder Urothelial Carcinoma', 'NCH'),
-        '5P': ('University Hospital Erlangen', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        '5Q': ('Proteogenex, Inc', 'Pancreatic adenocarcinoma', 'NCH'),
-        '5R': ('Proteogenex, Inc', 'Liver hepatocellular carcinoma', 'NCH'),
-        '5S': ('Holy Cross', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        '5T': ('Holy Cross', 'Breast invasive carcinoma', 'NCH'),
-        '5U': ('Regina Elena National Cancer Institute', 'Thymoma', 'NCH'),
-        '5V': ('Roswell Park', 'Thymoma', 'NCH'),
-        '5W': ('University of Alabama', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        '5X': ('University of Alabama', 'Ovarian serous cystadenocarcinoma', 'NCH'),
-        '60': ('Roswell Park', 'Lung squamous cell carcinoma', 'IGC'),
-        '61': ('University of Pittsburgh', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '62': ('Thoraxklinik at University Hospital Heidelberg', 'Lung adenocarcinoma', 'IGC'),
-        '63': ('Ontario Institute for Cancer Research', 'Lung squamous cell carcinoma', 'IGC'),
-        '64': ('Fox Chase', 'Lung adenocarcinoma', 'IGC'),
-        '65': ('Roswell Park', 'Glioblastoma multiforme', 'IGC'),
-        '66': ('Indivumed', 'Lung squamous cell carcinoma', 'IGC'),
-        '67': ('St Joseph\'s Medical Center (MD)', 'Lung adenocarcinoma', 'IGC'),
-        '68': ('Washington University - Cleveland Clinic', 'Lung squamous cell carcinoma', 'IGC'),
-        '69': ('Washington University - Cleveland Clinic', 'Lung adenocarcinoma', 'IGC'),
-        '6A': ('University of Kansas', 'Lung squamous cell carcinoma', 'NCH'),
-        '6D': ('University of Oklahoma HSC', 'Kidney renal clear cell carcinoma', 'NCH'),
-        '6G': ('University of Sao Paulo', 'Rectum adenocarcinoma', 'NCH'),
-        '70': ('ILSBio', 'Lung squamous cell carcinoma', 'IGC'),
-        '71': ('ILSBio', 'Lung adenocarcinoma', 'IGC'),
-        '72': ('NCH', 'Ovarian serous cystadenocarcinoma', 'IGC'),
-        '73': ('Roswell Park', 'Lung adenocarcinoma', 'IGC'),
-        '74': ('Swedish Neurosciences', 'Glioblastoma multiforme', 'IGC'),
-        '75': ('Ontario Institute for Cancer Research (OICR)', 'Lung adenocarcinoma', 'IGC'),
-        '76': ('Thomas Jefferson University', 'Glioblastoma multiforme', 'IGC'),
-        '77': ('Prince Charles Hospital', 'Lung squamous cell carcinoma', 'IGC'),
-        '78': ('Prince Charles Hospital', 'Lung adenocarcinoma', 'IGC'),
-        '79': ('Ontario Institute for Cancer Research (OICR)/Ottawa', 'Lung squamous cell carcinoma', 'IGC'),
-        '80': ('Ontario Institute for Cancer Research (OICR)/Ottawa', 'Lung adenocarcinoma', 'IGC'),
-        '81': ('CHI-Penrose Colorado', 'Glioblastoma multiforme', 'IGC'),
-        '82': ('CHI-Penrose Colorado', 'Lung squamous cell carcinoma', 'IGC'),
-        '83': ('CHI-Penrose Colorado', 'Lung adenocarcinoma', 'IGC'),
-        '85': ('Asterand', 'Lung squamous cell carcinoma', 'IGC'),
-        '86': ('Asterand', 'Lung adenocarcinoma', 'IGC'),
-        '87': ('International Genomics Consortium', 'Glioblastoma multiforme', 'IGC'),
-        '90': ('ABS - IUPUI', 'Lung squamous cell carcinoma', 'IGC'),
-        '91': ('ABS - IUPUI', 'Lung adenocarcinoma', 'IGC'),
-        '92': ('Washington University - St. Louis', 'Lung squamous cell carcinoma', 'IGC'),
-        '93': ('Washington University - St. Louis', 'Lung adenocarcinoma', 'IGC'),
-        '94': ('Washington University - Emory', 'Lung squamous cell carcinoma', 'IGC'),
-        '95': ('Washington University - Emory', 'Lung adenocarcinoma', 'IGC'),
-        '96': ('Washington University - NYU', 'Lung squamous cell carcinoma', 'IGC'),
-        '97': ('Washington University - NYU', 'Lung adenocarcinoma', 'IGC'),
-        '98': ('Washington University - Alabama', 'Lung squamous cell carcinoma', 'IGC'),
-        '99': ('Washington University - Alabama', 'Lung adenocarcinoma', 'IGC'),
-        'A1': ('UCSF', 'Breast invasive carcinoma', 'NCH'),
-        'A2': ('Walter Reed', 'Breast invasive carcinoma', 'NCH'),
-        'A3': ('International Genomics Consortium', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'A4': ('International Genomics Consortium', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'A5': ('Cedars Sinai', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'A6': ('Christiana Healthcare', 'Colon adenocarcinoma', 'IGC'),
-        'A7': ('Christiana Healthcare', 'Breast invasive carcinoma', 'NCH'),
-        'A8': ('Indivumed', 'Breast invasive carcinoma', 'NCH'),
-        'AA': ('Indivumed', 'Colon adenocarcinoma', 'IGC'),
-        'AB': ('Washington University', 'Acute Myeloid Leukemia', 'NCH'),
-        'AC': ('International Genomics Consortium', 'Breast invasive carcinoma', 'NCH'),
-        'AD': ('International Genomics Consortium', 'Colon adenocarcinoma', 'IGC'),
-        'AF': ('Christiana Healthcare', 'Rectum adenocarcinoma', 'IGC'),
-        'AG': ('Indivumed', 'Rectum adenocarcinoma', 'IGC'),
-        'AH': ('International Genomics Consortium', 'Rectum adenocarcinoma', 'IGC'),
-        'AJ': ('International Genomics Conosrtium', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'AK': ('Fox Chase', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'AL': ('Fox Chase', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'AM': ('Cureline', 'Colon adenocarcinoma', 'IGC'),
-        'AN': ('Cureline', 'Breast invasive carcinoma', 'NCH'),
-        'AO': ('MSKCC', 'Breast invasive carcinoma', 'NCH'),
-        'AP': ('MSKCC', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'AQ': ('UNC ', 'Breast invasive carcinoma', 'NCH'),
-        'AR': ('Mayo', 'Breast invasive carcinoma', 'NCH'),
-        'AS': ('St. Joseph\'s Medical Center-(MD)', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'AT': ('St. Joseph\'s Medical Center-(MD)', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'AU': ('St. Joseph\'s Medical Center-(MD)', 'Colon adenocarcinoma', 'IGC'),
-        'AV': ('NCH', 'Cell Line Control', 'NCH'),
-        'AW': ('Cureline', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'AX': ('Gynecologic Oncology Group', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'AY': ('UNC', 'Colon adenocarcinoma', 'IGC'),
-        'AZ': ('University of Pittsburgh', 'Colon adenocarcinoma', 'IGC'),
-        'B0': ('University of Pittsburgh', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'B1': ('University of Pittsburgh', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'B2': ('Christiana Healthcare', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'B3': ('Christiana Healthcare', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'B4': ('Cureline', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'B5': ('Duke', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'B6': ('Duke', 'Breast invasive carcinoma', 'NCH'),
-        'B7': ('Cureline', 'Stomach adenocarcinoma', 'IGC'),
-        'B8': ('UNC', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'B9': ('UNC', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'BA': ('UNC', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'BB': ('Johns Hopkins', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'BC': ('UNC', 'Liver hepatocellular carcinoma', 'NCH'),
-        'BD': ('University of Pittsburgh', 'Liver hepatocellular carcinoma', 'NCH'),
-        'BF': ('Cureline', 'Skin Cutaneous Melanoma', 'NCH'),
-        'BG': ('University of Pittsburgh', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'BH': ('University of Pittsburgh', 'Breast invasive carcinoma', 'NCH'),
-        'BI': ('University of Pittsburgh', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'BJ': ('University of Pittsburgh', 'Thyroid carcinoma', 'IGC'),
-        'BK': ('Christiana Healthcare', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'BL': ('Christiana Healthcare', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'BM': ('UNC', 'Rectum adenocarcinoma', 'IGC'),
-        'BP': ('MSKCC', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'BQ': ('MSKCC', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'BR': ('Asterand', 'Stomach adenocarcinoma', 'IGC'),
-        'BS': ('University of Hawaii', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'BT': ('University of Pittsburgh', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'BW': ('St. Joseph\'s Medical Center-(MD)', 'Liver hepatocellular carcinoma', 'NCH'),
-        'C4': ('Indivumed', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'C5': ('Medical College of Wisconsin', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'C8': ('ILSBio', 'Breast invasive carcinoma', 'NCH'),
-        'C9': ('ILSBio', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'CA': ('ILSBio', 'Colon adenocarcinoma', 'IGC'),
-        'CB': ('ILSBio', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'CC': ('ILSBio', 'Liver hepatocellular carcinoma', 'NCH'),
-        'CD': ('ILSBio', 'Stomach adenocarcinoma', 'IGC'),
-        'CE': ('ILSBio', 'Thyroid carcinoma', 'IGC'),
-        'CF': ('ILSBio', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'CG': ('Indivumed', 'Stomach adenocarcinoma', 'IGC'),
-        'CH': ('Indivumed', 'Prostate adenocarcinoma', 'IGC'),
-        'CI': ('University of Pittsburgh', 'Rectum adenocarcinoma', 'IGC'),
-        'CJ': ('MD Anderson Cancer Center', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'CK': ('Harvard', 'Colon adenocarcinoma', 'IGC'),
-        'CL': ('Harvard', 'Rectum adenocarcinoma', 'IGC'),
-        'CM': ('MSKCC', 'Colon adenocarcinoma', 'IGC'),
-        'CN': ('University of Pittsburgh', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'CQ': ('University Health Network, Toronto', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'CR': ('Vanderbilt University', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'CS': ('Thomas Jefferson University', 'Brain Lower Grade Glioma', 'IGC'),
-        'CU': ('UNC', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'CV': ('MD Anderson Cancer Center', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'CW': ('Mayo Clinic - Rochester', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'CX': ('Medical College of Georgia', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'CZ': ('Harvard', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'D1': ('Mayo Clinic', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'D3': ('MD Anderson', 'Skin Cutaneous Melanoma', 'NCH'),
-        'D5': ('Greater Poland Cancer Center', 'Colon adenocarcinoma', 'IGC'),
-        'D6': ('Greater Poland Cancer Center', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'D7': ('Greater Poland Cancer Center', 'Stomach adenocarcinoma', 'IGC'),
-        'D8': ('Greater Poland Cancer Center', 'Breast invasive carcinoma', 'NCH'),
-        'D9': ('Greater Poland Cancer Center', 'Skin Cutaneous Melanoma', 'NCH'),
-        'DA': ('Yale', 'Skin Cutaneous Melanoma', 'NCH'),
-        'DB': ('Mayo Clinic - Rochester', 'Brain Lower Grade Glioma', 'IGC'),
-        'DC': ('MSKCC', 'Rectum adenocarcinoma', 'IGC'),
-        'DD': ('Mayo Clinic - Rochester', 'Liver hepatocellular carcinoma', 'NCH'),
-        'DE': ('University of North Carolina', 'Thyroid carcinoma', 'NCH'),
-        'DF': ('Ontario Institute for Cancer Research', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'DG': ('Ontario Institute for Cancer Research', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'DH': ('University of Florida', 'Brain Lower Grade Glioma', 'IGC'),
-        'DI': ('MD Anderson', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'DJ': ('Memorial Sloan Kettering', 'Thyroid carcinoma', 'NCH'),
-        'DK': ('Memorial Sloan Kettering', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'DM': ('University Of Michigan', 'Colon adenocarcinoma', 'NCH'),
-        'DO': ('Medical College of Georgia', 'Thyroid carcinoma', 'NCH'),
-        'DQ': ('University Of Michigan', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'DR': ('University of Hawaii', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'DS': ('Cedars Sinai', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'DT': ('ILSBio', 'Rectum adenocarcinoma', 'IGC'),
-        'DU': ('Henry Ford Hospital', 'Brain Lower Grade Glioma', 'IGC'),
-        'DV': ('NCI Urologic Oncology Branch', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'DW': ('NCI Urologic Oncology Branch', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'DX': ('Memorial Sloan Kettering', 'Sarcoma', 'NCH'),
-        'DY': ('University Of Michigan', 'Rectum adenocarcinoma', 'NCH'),
-        'DZ': ('Mayo Clinic - Rochester', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'E1': ('Duke', 'Brain Lower Grade Glioma', 'IGC'),
-        'E2': ('Roswell Park', 'Breast invasive carcinoma', 'NCH'),
-        'E3': ('Roswell Park', 'Thyroid carcinoma', 'NCH'),
-        'E5': ('Roswell Park', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'E6': ('Roswell Park', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'E7': ('Asterand', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'E8': ('Asterand', 'Thyroid carcinoma', 'NCH'),
-        'E9': ('Asterand', 'Breast invasive carcinoma', 'NCH'),
-        'EA': ('Asterand', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'EB': ('Asterand', 'Skin Cutaneous Melanoma', 'NCH'),
-        'EC': ('Asterand', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'ED': ('Asterand', 'Liver hepatocellular carcinoma', 'NCH'),
-        'EE': ('University of Sydney', 'Skin Cutaneous Melanoma', 'NCH'),
-        'EF': ('Cureline', 'Rectum adenocarcinoma', 'IGC'),
-        'EI': ('Greater Poland Cancer Center', 'Rectum adenocarcinoma', 'IGC'),
-        'EJ': ('University of Pittsburgh', 'Prostate adenocarcinoma', 'IGC'),
-        'EK': ('Gynecologic Oncology Group', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'EL': ('MD Anderson', 'Thyroid carcinoma', 'NCH'),
-        'EM': ('University Health Network', 'Thyroid carcinoma', 'NCH'),
-        'EO': ('University Health Network', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'EP': ('Christiana Healthcare', 'Liver hepatocellular carcinoma', 'NCH'),
-        'EQ': ('Christiana Healthcare', 'Stomach adenocarcinoma', 'IGC'),
-        'ER': ('University of Pittsburgh', 'Skin Cutaneous Melanoma', 'NCH'),
-        'ES': ('University of Florida', 'Liver hepatocellular carcinoma', 'NCH'),
-        'ET': ('Johns Hopkins', 'Thyroid carcinoma', 'NCH'),
-        'EU': ('CHI-Penrose Colorado', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'EV': ('CHI-Penrose Colorado', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'EW': ('University of Miami', 'Breast invasive carcinoma', 'NCH'),
-        'EX': ('University of North Carolina', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'EY': ('University of North Carolina', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'EZ': ('UNC', 'Brain Lower Grade Glioma', 'IGC'),
-        'F1': ('UNC', 'Stomach adenocarcinoma', 'IGC'),
-        'F2': ('UNC', 'Pancreatic adenocarcinoma', 'IGC'),
-        'F4': ('Asterand', 'Colon adenocarcinoma', 'IGC'),
-        'F5': ('Asterand', 'Rectum adenocarcinoma', 'IGC'),
-        'F6': ('Asterand', 'Brain Lower Grade Glioma', 'IGC'),
-        'F7': ('Asterand', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'F9': ('Asterand', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'FA': ('Asterand', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'IGC'),
-        'FB': ('Asterand', 'Pancreatic adenocarcinoma', 'IGC'),
-        'FC': ('Asterand', 'Prostate adenocarcinoma', 'IGC'),
-        'FD': ('BLN - University Of Chicago', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'FE': ('Ohio State University', 'Thyroid carcinoma', 'NCH'),
-        'FF': ('SingHealth', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'IGC'),
-        'FG': ('Case Western', 'Brain Lower Grade Glioma', 'IGC'),
-        'FH': ('CHI-Penrose Colorado', 'Thyroid carcinoma', 'NCH'),
-        'FI': ('Washington University', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'FJ': ('BLN - Baylor', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'FK': ('Johns Hopkins', 'Thyroid carcinoma', 'NCH'),
-        'FL': ('University of Hawaii - Normal Study', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'FM': ('International Genomics Consortium', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'IGC'),
-        'FN': ('International Genomics Consortium', 'Brain Lower Grade Glioma', 'IGC'),
-        'FP': ('International Genomics Consortium', 'Stomach adenocarcinoma', 'IGC'),
-        'FQ': ('Johns Hopkins', 'Pancreatic adenocarcinoma', 'IGC'),
-        'FR': ('University of North Carolina', 'Skin Cutaneous Melanoma', 'NCH'),
-        'FS': ('Essen', 'Skin Cutaneous Melanoma', 'NCH'),
-        'FT': ('BLN - University of Miami', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'FU': ('International Genomics Consortium', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'FV': ('International Genomics Consortium', 'Liver hepatocellular carcinoma', 'NCH'),
-        'FW': ('International Genomics Consortium', 'Skin Cutaneous Melanoma', 'NCH'),
-        'FX': ('International Genomics Consortium', 'Sarcoma', 'NCH'),
-        'FY': ('International Genomics Consortium', 'Thyroid carcinoma', 'NCH'),
-        'FZ': ('University of Pittsburgh', 'Pancreatic adenocarcinoma', 'IGC'),
-        'G2': ('MD Anderson', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'G3': ('Alberta Health Services', 'Liver hepatocellular carcinoma', 'NCH'),
-        'G4': ('Roswell Park', 'Colon adenocarcinoma', 'IGC'),
-        'G5': ('Roswell Park', 'Rectum adenocarcinoma', 'IGC'),
-        'G6': ('Roswell Park', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'G7': ('Roswell Park', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'G8': ('Roswell Park', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'IGC'),
-        'G9': ('Roswell Park', 'Prostate adenocarcinoma', 'IGC'),
-        'GC': ('International Genomics Consortium', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'GD': ('ABS - IUPUI', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'GE': ('ABS - IUPUI', 'Thyroid carcinoma', 'NCH'),
-        'GF': ('ABS - IUPUI', 'Skin Cutaneous Melanoma', 'NCH'),
-        'GG': ('ABS - IUPUI', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'GH': ('ABS - IUPUI', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'GI': ('ABS - IUPUI', 'Breast invasive carcinoma', 'NCH'),
-        'GJ': ('ABS - IUPUI', 'Liver hepatocellular carcinoma', 'NCH'),
-        'GK': ('ABS - IUPUI', 'Kidney renal clear cell carcinoma', 'IGC'),
-        'GL': ('ABS - IUPUI', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'GM': ('MD Anderson', 'Breast invasive carcinoma', 'NCH'),
-        'GN': ('Roswell', 'Skin Cutaneous Melanoma', 'NCH'),
-        'GP': ('MD Anderson', 'Acute Myeloid Leukemia', 'NCH'),
-        'GR': ('University of Nebraska Medical Center (UNMC)', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'IGC'),
-        'GS': ('Fundacio Clinic per a la Recerca Biomedica', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'IGC'),
-        'GU': ('BLN - UT Southwestern Medical Center at Dallas', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'GV': ('BLN - Cleveland Clinic', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'GZ': ('BC Cancer Agency', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'IGC'),
-        'H1': ('Medical College of Georgia', 'Stomach adenocarcinoma', 'IGC'),
-        'H2': ('Christiana Healthcare', 'Thyroid carcinoma', 'NCH'),
-        'H3': ('ABS - IUPUI', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'IGC'),
-        'H4': ('Medical College of Georgia', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'H5': ('Medical College of Georgia', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'H6': ('Christiana Healthcare', 'Pancreatic adenocarcinoma', 'IGC'),
-        'H7': ('ABS - IUPUI', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'H8': ('ABS - IUPUI', 'Pancreatic adenocarcinoma', 'IGC'),
-        'H9': ('ABS - IUPUI', 'Prostate adenocarcinoma', 'IGC'),
-        'HA': ('Alberta Health Services', 'Stomach adenocarcinoma', 'IGC'),
-        'HB': ('University of North Carolina', 'Sarcoma', 'NCH'),
-        'HC': ('International Genomics Consortium', 'Prostate adenocarcinoma', 'IGC'),
-        'HD': ('International Genomics Consortium', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'HE': ('Ontario Institute for Cancer Research (OICR)', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'HF': ('Ontario Institute for Cancer Research (OICR)', 'Stomach adenocarcinoma', 'IGC'),
-        'HG': ('Roswell Park', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'HH': ('Fox Chase', 'Stomach adenocarcinoma', 'IGC'),
-        'HI': ('Fox Chase', 'Prostate adenocarcinoma', 'IGC'),
-        'HJ': ('Fox Chase', 'Stomach adenocarcinoma', 'IGC'),
-        'HK': ('Fox Chase', 'Brain Lower Grade Glioma', 'IGC'),
-        'HL': ('Fox Chase', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'HM': ('Christiana Healthcare', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'HN': ('Ontario Institute for Cancer Research (OICR)', 'Breast invasive carcinoma', 'NCH'),
-        'HP': ('Ontario Institute for Cancer Research (OICR)', 'Liver hepatocellular carcinoma', 'NCH'),
-        'HQ': ('Ontario Institute for Cancer Research (OICR)', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'HR': ('Ontario Institute for Cancer Research (OICR)', 'Skin Cutaneous Melanoma', 'NCH'),
-        'HS': ('Ontario Institute for Cancer Research (OICR)', 'Sarcoma', 'NCH'),
-        'HT': ('Case Western - St Joes', 'Brain Lower Grade Glioma', 'IGC'),
-        'HU': ('National Cancer Center Korea', 'Stomach adenocarcinoma', 'IGC'),
-        'HV': ('National Cancer Center Korea', 'Pancreatic adenocarcinoma', 'IGC'),
-        'HW': ('MSKCC', 'Brain Lower Grade Glioma', 'IGC'),
-        'HZ': ('International Genomics Consortium', 'Pancreatic adenocarcinoma', 'IGC'),
-        'IA': ('Cleveland Clinic', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'IB': ('Alberta Health Services', 'Pancreatic adenocarcinoma', 'IGC'),
-        'IC': ('University of Pittsburgh', 'Esophageal carcinoma ', 'NCH'),
-        'IE': ('ABS - IUPUI', 'Sarcoma', 'NCH'),
-        'IF': ('University of Texas MD Anderson Cancer Center', 'Sarcoma', 'NCH'),
-        'IG': ('Asterand', 'Esophageal carcinoma ', 'NCH'),
-        'IH': ('University of Miami', 'Skin Cutaneous Melanoma', 'NCH'),
-        'IJ': ('Christiana Healthcare', 'Acute Myeloid Leukemia', 'NCH'),
-        'IK': ('Christiana Healthcare', 'Brain Lower Grade Glioma', 'IGC'),
-        'IM': ('University of Miami', 'Thyroid carcinoma', 'NCH'),
-        'IN': ('University of Pittsburgh', 'Stomach adenocarcinoma', 'IGC'),
-        'IP': ('ABS - IUPUI', 'Stomach adenocarcinoma', 'IGC'),
-        'IQ': ('University of Miami', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'IR': ('Memorial Sloan Kettering', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'IS': ('Memorial Sloan Kettering', 'Sarcoma', 'NCH'),
-        'IW': ('Cedars Sinai', 'Sarcoma', 'NCH'),
-        'IZ': ('ABS - Lahey Clinic', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'J1': ('ABS - Lahey Clinic', 'Lung squamous cell carcinoma', 'IGC'),
-        'J2': ('ABS - Lahey Clinic', 'Lung adenocarcinoma', 'IGC'),
-        'J4': ('ABS - Lahey Clinic', 'Prostate adenocarcinoma', 'IGC'),
-        'J7': ('ILSBio', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'J8': ('Mayo Clinic', 'Thyroid carcinoma', 'NCH'),
-        'J9': ('Melbourne Health', 'Prostate adenocarcinoma', 'IGC'),
-        'JA': ('ABS - Research Metrics Pakistan', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'JL': ('ABS - Research Metrics Pakistan', 'Breast invasive carcinoma', 'NCH'),
-        'JU': ('BLN - Baylor', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'JV': ('BLN - Baylor', 'Sarcoma', 'NCH'),
-        'JW': ('BLN - Baylor', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'JX': ('Washington University', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'JY': ('University Health Network', 'Esophageal carcinoma ', 'NCH'),
-        'JZ': ('University of Rochester', 'Esophageal carcinoma ', 'NCH'),
-        'K1': ('University of Pittsburgh', 'Sarcoma', 'NCH'),
-        'K4': ('ABS - Lahey Clinic', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'K6': ('ABS - Lahey Clinic', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'K7': ('ABS - Lahey Clinic', 'Liver hepatocellular carcinoma', 'NCH'),
-        'K8': ('ABS - Lahey Clinic', 'Skin Cutaneous Melanoma', 'NCH'),
-        'KA': ('ABS - Lahey Clinic', 'Esophageal carcinoma ', 'NCH'),
-        'KB': ('University Health Network, Toronto', 'Stomach adenocarcinoma', 'IGC'),
-        'KC': ('Cornell Medical College', 'Prostate adenocarcinoma', 'IGC'),
-        'KD': ('Mount Sinai School of Medicine', 'Sarcoma', 'NCH'),
-        'KE': ('Mount Sinai School of Medicine', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'KF': ('Christiana Healthcare', 'Sarcoma', 'NCH'),
-        'KG': ('Baylor Network', 'Pancreatic adenocarcinoma', 'IGC'),
-        'KH': ('Memorial Sloan Kettering', 'Esophageal carcinoma ', 'NCH'),
-        'KJ': ('University of Miami', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'KK': ('MD Anderson Cancer Center', 'Prostate adenocarcinoma', 'IGC'),
-        'KL': ('MSKCC', 'Kidney Chromophobe', 'IGC'),
-        'KM': ('NCI Urologic Oncology Branch', 'Kidney Chromophobe', 'IGC'),
-        'KN': ('Harvard', 'Kidney Chromophobe', 'IGC'),
-        'KO': ('MD Anderson Cancer Center', 'Kidney Chromophobe', 'IGC'),
-        'KP': ('British Columbia Cancer Agency', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'KQ': ('Cornell Medical College', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'KR': ('University Of Michigan', 'Liver hepatocellular carcinoma', 'NCH'),
-        'KS': ('University Of Michigan', 'Thyroid carcinoma', 'NCH'),
-        'KT': ('Hartford', 'Brain Lower Grade Glioma', 'IGC'),
-        'KU': ('Hartford', 'Head and Neck squamous cell carcinoma', 'IGC'),
-        'KV': ('Hartford', 'Kidney renal papillary cell carcinoma', 'IGC'),
-        'KZ': ('Hartford', 'Stomach adenocarcinoma', 'IGC'),
-        'L1': ('Hartford', 'Pancreatic adenocarcinoma', 'IGC'),
-        'L3': ('Gundersen Lutheran Health System', 'Lung squamous cell carcinoma', 'IGC'),
-        'L4': ('Gundersen Lutheran Health System', 'Lung adenocarcinoma', 'IGC'),
-        'L5': ('University of Michigan', 'Esophageal carcinoma ', 'NCH'),
-        'L6': ('National Institutes of Health', 'Thyroid carcinoma', 'NCH'),
-        'L7': ('Christiana Care', 'Esophageal carcinoma ', 'NCH'),
-        'L8': ('University of Miami', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'L9': ('Candler', 'Lung adenocarcinoma', 'IGC'),
-        'LA': ('Candler', 'Lung squamous cell carcinoma', 'IGC'),
-        'LB': ('Candler', 'Pancreatic adenocarcinoma', 'IGC'),
-        'LC': ('Hartford Hospital', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'LD': ('Hartford Hospital', 'Breast invasive carcinoma', 'NCH'),
-        'LG': ('Hartford Hospital', 'Liver hepatocellular carcinoma', 'NCH'),
-        'LH': ('Hartford Hospital', 'Skin Cutaneous Melanoma', 'NCH'),
-        'LI': ('Hartford Hospital', 'Sarcoma', 'NCH'),
-        'LK': ('University of Pittsburgh', 'Mesothelioma', 'NCH'),
-        'LL': ('Candler', 'Breast invasive carcinoma', 'NCH'),
-        'LN': ('ILSBIO', 'Esophageal carcinoma ', 'NCH'),
-        'LP': ('ILSBIO', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'LQ': ('Gundersen Lutheran Health System', 'Breast invasive carcinoma', 'NCH'),
-        'LS': ('Gundersen Lutheran Health System', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'LT': ('Gundersen Lutheran Health System', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'M7': ('University of North Carolina', 'Prostate adenocarcinoma', 'NCH'),
-        'M8': ('Ontario Institute for Cancer Research (OICR)', 'Pancreatic adenocarcinoma', 'NCH'),
-        'M9': ('Ontario Institute for Cancer Research (OICR)', 'Esophageal carcinoma ', 'NCH'),
-        'MA': ('MD Anderson Cancer Center', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'MB': ('University of Minnesota', 'Sarcoma', 'NCH'),
-        'ME': ('University of Minnesota', 'Lung adenocarcinoma', 'NCH'),
-        'MF': ('University of Minnesota', 'Lung squamous cell carcinoma', 'NCH'),
-        'MG': ('BLN - Baylor', 'Prostate adenocarcinoma', 'NCH'),
-        'MH': ('BLN - Baylor', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'MI': ('BLN - Baylor', 'Liver hepatocellular carcinoma', 'NCH'),
-        'MJ': ('BLN - Baylor', 'Sarcoma', 'NCH'),
-        'MK': ('BLN - Baylor', 'Thyroid carcinoma', 'NCH'),
-        'ML': ('BLN - Baylor', 'Lung squamous cell carcinoma', 'NCH'),
-        'MM': ('BLN - Baylor', 'Kidney renal clear cell carcinoma', 'NCH'),
-        'MN': ('BLN - Baylor', 'Lung adenocarcinoma', 'NCH'),
-        'MO': ('ILSBio', 'Sarcoma', 'NCH'),
-        'MP': ('Washington University - Mayo Clinic', 'Lung adenocarcinoma', 'NCH'),
-        'MQ': ('Washington University - NYU', 'Mesothelioma', 'NCH'),
-        'MR': ('University of Minnesota', 'Liver hepatocellular carcinoma', 'NCH'),
-        'MS': ('University of Minnesota', 'Breast invasive carcinoma', 'NCH'),
-        'MT': ('University of Minnesota', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'MU': ('University of Minnesota', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'MV': ('University of Minnesota', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'MW': ('University of Miami', 'Kidney renal clear cell carcinoma', 'NCH'),
-        'MX': ('MSKCC', 'Stomach adenocarcinoma', 'NCH'),
-        'MY': ('Montefiore Medical Center', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'MZ': ('Montefiore Medical Center', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'N1': ('Montefiore Medical Center', 'Sarcoma', 'NCH'),
-        'N5': ('MSKCC', 'Uterine Carcinosarcoma', 'NCH'),
-        'N6': ('University of Pittsburgh', 'Uterine Carcinosarcoma', 'NCH'),
-        'N7': ('Washington University', 'Uterine Carcinosarcoma', 'NCH'),
-        'N8': ('University of North Carolina', 'Uterine Carcinosarcoma', 'NCH'),
-        'N9': ('MD Anderson', 'Uterine Carcinosarcoma', 'NCH'),
-        'NA': ('Duke University', 'Uterine Carcinosarcoma', 'NCH'),
-        'NB': ('Washington University - CHUV', 'Lung adenocarcinoma', 'NCH'),
-        'NC': ('Washington University - CHUV', 'Lung squamous cell carcinoma', 'NCH'),
-        'ND': ('Cedars Sinai', 'Uterine Carcinosarcoma', 'NCH'),
-        'NF': ('Mayo Clinic - Rochester', 'Uterine Carcinosarcoma', 'NCH'),
-        'NG': ('Roswell Park', 'Uterine Carcinosarcoma', 'NCH'),
-        'NH': ('Candler', 'Colon adenocarcinoma', 'NCH'),
-        'NI': ('Roswell Park', 'Liver hepatocellular carcinoma', 'NCH'),
-        'NJ': ('Washington University - Rush University', 'Lung adenocarcinoma', 'NCH'),
-        'NK': ('Washington University - Rush University', 'Lung squamous cell carcinoma', 'NCH'),
-        'NM': ('Cambridge BioSource', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'NP': ('International Genomics Consortium', 'Kidney Chromophobe', 'NCH'),
-        'NQ': ('International Genomics Consortium', 'Mesothelioma', 'NCH'),
-        'NS': ('Gundersen Lutheran Health System', 'Skin Cutaneous Melanoma', 'NCH'),
-        'O1': ('Washington University - CALGB', 'Lung adenocarcinoma', 'NCH'),
-        'O2': ('Washington University - CALGB', 'Lung squamous cell carcinoma', 'NCH'),
-        'O8': ('Saint Mary\'s Health Care', 'Liver hepatocellular carcinoma', 'NCH'),
-        'O9': ('Saint Mary\'s Health Care', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'OC': ('Saint Mary\'s Health Care', 'Lung squamous cell carcinoma', 'NCH'),
-        'OD': ('Saint Mary\'s Health Care', 'Skin Cutaneous Melanoma', 'NCH'),
-        'OE': ('Saint Mary\'s Health Care', 'Pancreatic adenocarcinoma', 'NCH'),
-        'OJ': ('Saint Mary\'s Health Care', 'Thyroid carcinoma', 'NCH'),
-        'OK': ('Mount Sinai School of Medicine', 'Breast invasive carcinoma', 'NCH'),
-        'OL': ('University of Chicago', 'Breast invasive carcinoma', 'NCH'),
-        'OR': ('University of Michigan', 'Adrenocortical carcinoma', 'NCH'),
-        'OU': ('Roswell Park', 'Adrenocortical carcinoma', 'NCH'),
-        'OW': ('International Genomics Consortium', 'Miscellaneous', 'NCH'),
-        'OX': ('University of North Carolina', 'Glioblastoma multiforme', 'NCH'),
-        'OY': ('University of North Carolina', 'Ovarian serous cystadenocarcinoma', 'NCH'),
-        'P3': ('Fred Hutchinson', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'P4': ('MD Anderson Cancer Center', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'P5': ('Cureline', 'Brain Lower Grade Glioma', 'NCH'),
-        'P6': ('Translational Genomics Research Institute', 'Adrenocortical carcinoma', 'NCH'),
-        'P7': ('Translational Genomics Research Institute', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'P8': ('University of Pittsburgh', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'P9': ('University of Minnesota', 'Pancreatic adenocarcinoma', 'NCH'),
-        'PA': ('University of Minnesota', 'Adrenocortical carcinoma', 'NCH'),
-        'PB': ('University of Minnesota', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'NCH'),
-        'PC': ('Fox Chase', 'Sarcoma', 'NCH'),
-        'PD': ('Fox Chase', 'Liver hepatocellular carcinoma', 'NCH'),
-        'PE': ('Fox Chase', 'Breast invasive carcinoma', 'NCH'),
-        'PG': ('Montefiore Medical Center', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'PH': ('Gundersen Lutheran', 'Acute Myeloid Leukemia', 'NCH'),
-        'PJ': ('Gundersen Lutheran', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'PK': ('University Health Network', 'Adrenocortical carcinoma', 'NCH'),
-        'PL': ('Institute of Human Virology Nigeria', 'Breast invasive carcinoma', 'NCH'),
-        'PN': ('Institute of Human Virology Nigeria', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'PQ': ('University of Colorado Denver', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'PR': ('Roswell Park', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'PT': ('Maine Medical Center', 'Sarcoma', 'NCH'),
-        'PZ': ('ABS - Lahey Clinic', 'Pancreatic adenocarcinoma', 'NCH'),
-        'Q1': ('University of Oklahoma HSC', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'Q2': ('University of Oklahoma HSC', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'Q3': ('University of Oklahoma HSC', 'Pancreatic adenocarcinoma', 'NCH'),
-        'Q4': ('Emory University', 'Acute Myeloid Leukemia', 'NCH'),
-        'Q9': ('Emory University', 'Esophageal carcinoma ', 'NCH'),
-        'QA': ('Emory University', 'Liver hepatocellular carcinoma', 'NCH'),
-        'QB': ('Emory University', 'Skin Cutaneous Melanoma', 'NCH'),
-        'QC': ('Emory University', 'Sarcoma', 'NCH'),
-        'QD': ('Emory University', 'Thyroid carcinoma', 'NCH'),
-        'QF': ('BLN - Baylor', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'QG': ('BLN - Baylor', 'Colon adenocarcinoma', 'NCH'),
-        'QH': ('Fondazione-Besta', 'Brain Lower Grade Glioma', 'NCH'),
-        'QJ': ('Mount Sinai School of Medicine', 'Ovarian serous cystadenocarcinoma', 'NCH'),
-        'QK': ('Emory University - Winship Cancer Inst.', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'QL': ('University of Chicago', 'Colon adenocarcinoma', 'NCH'),
-        'QM': ('University of Oklahoma HSC', 'Uterine Carcinosarcoma', 'NCH'),
-        'QN': ('ILSBio', 'Uterine Carcinosarcoma', 'NCH'),
-        'QQ': ('Roswell Park', 'Sarcoma', 'NCH'),
-        'QR': ('National Institutes of Health', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'QS': ('Candler', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'QT': ('University of North Carolina', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'QU': ('Harvard Beth Israel', 'Prostate adenocarcinoma', 'NCH'),
-        'QV': ('Instituto Nacional de Cancerologia', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'QW': ('Instituto Nacional de Cancerologia', 'Stomach adenocarcinoma', 'NCH'),
-        'R1': ('CHI-Penrose Colorado', 'Colon adenocarcinoma', 'NCH'),
-        'R2': ('CHI-Penrose Colorado', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'R3': ('CHI-Penrose Colorado', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'R5': ('MD Anderson Cancer Center', 'Stomach adenocarcinoma', 'NCH'),
-        'R6': ('MD Anderson Cancer Center', 'Esophageal carcinoma ', 'NCH'),
-        'R7': ('Gundersen Lutheran Health System', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'R8': ('MD Anderson', 'Brain Lower Grade Glioma', 'NCH'),
-        'R9': ('Candler', 'Ovarian serous cystadenocarcinoma', 'NCH'),
-        'RA': ('Candler', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'RB': ('Emory University', 'Pancreatic adenocarcinoma', 'NCH'),
-        'RC': ('University of Utah', 'Liver hepatocellular carcinoma', 'NCH'),
-        'RD': ('Peter MacCallum Cancer Center', 'Stomach adenocarcinoma', 'NCH'),
-        'RE': ('Peter MacCallum Cancer Center', 'Esophageal carcinoma ', 'NCH'),
-        'RG': ('Montefiore Medical Center', 'Liver hepatocellular carcinoma', 'NCH'),
-        'RH': ('BLN - Baylor', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'RL': ('St. Joseph\'s Hospital AZ', 'Pancreatic adenocarcinoma', 'NCH'),
-        'RM': ('St. Joseph\'s Hospital AZ', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'RN': ('St. Joseph\'s Hospital AZ', 'Sarcoma', 'NCH'),
-        'RP': ('St. Joseph\'s Hospital AZ', 'Skin Cutaneous Melanoma', 'NCH'),
-        'RQ': ('St. Joseph\'s Hospital AZ', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'NCH'),
-        'RR': ('St. Joseph\'s Hospital AZ', 'Glioblastoma multiforme', 'NCH'),
-        'RS': ('Memorial Sloan Kettering Cancer Center', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'RT': ('Cleveland Clinic Foundation', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'RU': ('Northwestern University', 'Colon adenocarcinoma', 'NCH'),
-        'RV': ('Northwestern University', 'Pancreatic adenocarcinoma', 'NCH'),
-        'RW': ('Michigan University', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'RX': ('University of Minnesota', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'RY': ('University of California San Francisco', 'Brain Lower Grade Glioma', 'NCH'),
-        'RZ': ('Wills Eye Institute', 'Uveal Melanoma', 'NCH'),
-        'S2': ('Albert Einstein Medical Center', 'Lung adenocarcinoma', 'NCH'),
-        'S3': ('Albert Einstein Medical Center', 'Breast invasive carcinoma', 'NCH'),
-        'S4': ('University of Chicago', 'Pancreatic adenocarcinoma', 'NCH'),
-        'S5': ('University of Oklahoma HSC', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'S6': ('Gundersen Lutheran Health System', 'Testicular Germ Cell Tumors', 'NCH'),
-        'S7': ('University Hospital Motol', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'S8': ('ABS - IUPUI', 'Esophageal carcinoma ', 'NCH'),
-        'S9': ('Dept of Neurosurgery at University of Heidelberg', 'Brain Lower Grade Glioma', 'NCH'),
-        'SA': ('ABS - IUPUI', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'SB': ('Baylor College of Medicine', 'Testicular Germ Cell Tumors', 'NCH'),
-        'SC': ('Memorial Sloan Kettering', 'Mesothelioma', 'NCH'),
-        'SD': ('MD Anderson', 'Pancreatic adenocarcinoma', 'NCH'),
-        'SE': ('Boston Medical Center', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'SG': ('Cleveland Clinic Foundation', 'Sarcoma', 'NCH'),
-        'SH': ('Papworth Hospital', 'Mesothelioma', 'NCH'),
-        'SI': ('Washington University St. Louis', 'Sarcoma', 'NCH'),
-        'SJ': ('Albert Einstein Medical Center', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'SK': ('St. Joseph\'s Hospital AZ', 'Colon adenocarcinoma', 'NCH'),
-        'SL': ('St. Joseph\'s Hospital AZ', 'Uterine Corpus Endometrial Carcinoma', 'NCH'),
-        'SN': ('BLN - Baylor', 'Testicular Germ Cell Tumors', 'NCH'),
-        'SO': ('University of Minnesota', 'Testicular Germ Cell Tumors', 'NCH'),
-        'SP': ('University Health Network', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'SQ': ('International Genomics Consortium', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'SR': ('Tufts Medical Center', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'SS': ('Medical College of Georgia', 'Colon adenocarcinoma', 'NCH'),
-        'ST': ('Global Bioclinical-Moldova', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'SU': ('Global Bioclinical-Moldova', 'Prostate adenocarcinoma', 'NCH'),
-        'SW': ('Global Bioclinical-Moldova', 'Stomach adenocarcinoma', 'NCH'),
-        'SX': ('Mayo Clinic Arizona', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'SY': ('Mayo Clinic Arizona', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'T1': ('St. Joseph\'s Hospital Arizona', 'Liver hepatocellular carcinoma', 'NCH'),
-        'T2': ('St. University of Colorado Denver', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'T3': ('Molecular Response', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'T6': ('Molecular Response', 'Lung adenocarcinoma', 'NCH'),
-        'T7': ('Molecular Response', 'Kidney renal clear cell carcinoma', 'NCH'),
-        'T9': ('Molecular Response', 'Colon adenocarcinoma', 'NCH'),
-        'TE': ('Global BioClinical - Georgia', 'Skin Cutaneous Melanoma', 'NCH'),
-        'TG': ('Global BioClinical - Georgia', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'TK': ('Global BioClinical - Georgia', 'Prostate adenocarcinoma', 'NCH'),
-        'TL': ('Global BioClinical - Georgia', 'Stomach adenocarcinoma', 'NCH'),
-        'TM': ('The University of New South Wales', 'Brain Lower Grade Glioma', 'NCH'),
-        'TN': ('Ohio State University', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'TP': ('Maine Medical Center', 'Prostate adenocarcinoma', 'NCH'),
-        'TQ': ('University of Sao Paulo', 'Brain Lower Grade Glioma', 'NCH'),
-        'TR': ('Global Bioclinical-Moldova', 'Skin Cutaneous Melanoma', 'NCH'),
-        'TS': ('University of Pennsylvania', 'Mesothelioma', 'NCH'),
-        'TT': ('University of Pennsylvania', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'TV': ('Wake Forest University', 'Breast invasive carcinoma', 'NCH'),
-        'UB': ('UCSF', 'Liver hepatocellular carcinoma', 'NCH'),
-        'UC': ('University of Washington', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'UD': ('University of Western Australia', 'Mesothelioma', 'NCH'),
-        'UE': ('Asterand', 'Sarcoma', 'NCH'),
-        'UF': ('Barretos Cancer Hospital', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'UJ': ('Boston Medical Center', 'Lung squamous cell carcinoma', 'NCH'),
-        'UL': ('Boston Medical Center', 'Breast invasive carcinoma', 'NCH'),
-        'UN': ('Boston Medical Center', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'UP': ('Boston Medical Center', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'UR': ('Boston Medical Center', 'Prostate adenocarcinoma', 'NCH'),
-        'US': ('Garvan Institute of Medical Research', 'Pancreatic adenocarcinoma', 'NCH'),
-        'UT': ('Asbestos Diseases Research Institute', 'Mesothelioma', 'NCH'),
-        'UU': ('Mary Bird Perkins Cancer Center - Our Lady of the Lake', 'Breast invasive carcinoma', 'NCH'),
-        'UV': ('Capital Biosciences', 'Liver hepatocellular carcinoma', 'NCH'),
-        'UW': ('University of North Carolina', 'Kidney Chromophobe', 'NCH'),
-        'UY': ('University of California San Francisco', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'UZ': ('University of California San Francisco', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'V1': ('University of California San Francisco', 'Prostate adenocarcinoma', 'NCH'),
-        'V2': ('Cleveland Clinic Foundation', 'Prostate adenocarcinoma', 'NCH'),
-        'V3': ('Cleveland Clinic Foundation', 'Uveal Melanoma', 'NCH'),
-        'V4': ('Institut Curie', 'Uveal Melanoma', 'NCH'),
-        'V5': ('Duke University', 'Esophageal carcinoma ', 'NCH'),
-        'V6': ('Duke University', 'Stomach adenocarcinoma', 'NCH'),
-        'V7': ('Medical College of Georgia', 'Breast invasive carcinoma', 'NCH'),
-        'V8': ('Medical College of Georgia', 'Kidney renal clear cell carcinoma', 'NCH'),
-        'V9': ('Medical College of Georgia', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'VA': ('Alliance', 'Stomach adenocarcinoma', 'NCH'),
-        'VB': ('Global BioClinical - Georgia', 'Lymphoid Neoplasm Diffuse Large B-cell Lymphoma', 'NCH'),
-        'VD': ('University of Liverpool', 'Uveal Melanoma', 'NCH'),
-        'VF': ('University of Pennsylvania', 'Testicular Germ Cell Tumors', 'NCH'),
-        'VG': ('Institute of Human Virology Nigeria', 'Ovarian serous cystadenocarcinoma', 'NCH'),
-        'VK': ('Institute of Human Virology Nigeria', 'Colon adenocarcinoma', 'NCH'),
-        'VL': ('Institute of Human Virology Nigeria', 'Rectum adenocarcinoma', 'NCH'),
-        'VM': ('Huntsman Cancer Institute', 'Brain Lower Grade Glioma', 'NCH'),
-        'VN': ('NCI Urologic Oncology Branch', 'Prostate adenocarcinoma', 'NCH'),
-        'VP': ('Washington University', 'Prostate adenocarcinoma', 'NCH'),
-        'VQ': ('Barretos Cancer Hospital', 'Stomach adenocarcinoma', 'NCH'),
-        'VR': ('Barretos Cancer Hospital', 'Esophageal carcinoma ', 'NCH'),
-        'VS': ('Barretos Cancer Hospital', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'VT': ('Vanderbilt', 'Sarcoma', 'NCH'),
-        'VV': ('John Wayne Cancer Center', 'Brain Lower Grade Glioma', 'NCH'),
-        'VW': ('Northwestern University', 'Brain Lower Grade Glioma', 'NCH'),
-        'VX': ('Northwestern University', 'Stomach adenocarcinoma', 'NCH'),
-        'VZ': ('Albert Einstein Medical Center', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'W2': ('Medical College of Wisconsin', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'W3': ('John Wayne Cancer Center', 'Skin Cutaneous Melanoma', 'NCH'),
-        'W4': ('University of North Carolina', 'Testicular Germ Cell Tumors', 'NCH'),
-        'W5': ('Mayo Clinic Rochester', 'Cholangiocarcinoma', 'NCH'),
-        'W6': ('UCSF', 'Cholangiocarcinoma', 'NCH'),
-        'W7': ('Garvan Institute of Medical Research', 'Cholangiocarcinoma', 'NCH'),
-        'W8': ('Greenville Health System', 'Breast invasive carcinoma', 'NCH'),
-        'W9': ('University of Kansas', 'Brain Lower Grade Glioma', 'NCH'),
-        'WA': ('University of Schleswig-Holstein', 'Head and Neck squamous cell carcinoma', 'NCH'),
-        'WB': ('Erasmus MC', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'WC': ('MD Anderson', 'Uveal Melanoma', 'NCH'),
-        'WD': ('Emory University', 'Cholangiocarcinoma', 'NCH'),
-        'WE': ('Norfolk and Norwich Hospital', 'Skin Cutaneous Melanoma', 'NCH'),
-        'WF': ('Greenville Health System', 'Pancreatic adenocarcinoma', 'NCH'),
-        'WG': ('Greenville Health System', 'Lung squamous cell carcinoma', 'NCH'),
-        'WH': ('Greenville Health System', 'Brain Lower Grade Glioma', 'NCH'),
-        'WJ': ('Greenville Health System', 'Liver hepatocellular carcinoma', 'NCH'),
-        'WK': ('Brigham and Women\'s Hospital', 'Sarcoma', 'NCH'),
-        'WL': ('University of Kansas', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'WM': ('University of Kansas', 'Kidney renal clear cell carcinoma', 'NCH'),
-        'WN': ('University of Kansas', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'WP': ('University of Kansas', 'Sarcoma', 'NCH'),
-        'WQ': ('University of Kansas', 'Liver hepatocellular carcinoma', 'NCH'),
-        'WR': ('University of Kansas', 'Ovarian serous cystadenocarcinoma', 'NCH'),
-        'WS': ('University of Kansas', 'Colon adenocarcinoma', 'NCH'),
-        'WT': ('University of Kansas', 'Breast invasive carcinoma', 'NCH'),
-        'WU': ('Wake Forest University', 'Colon adenocarcinoma', 'NCH'),
-        'WW': ('Wake Forest University', 'Prostate adenocarcinoma', 'NCH'),
-        'WX': ('Yale University', 'Liver hepatocellular carcinoma', 'NCH'),
-        'WY': ('Johns Hopkins', 'Brain Lower Grade Glioma', 'NCH'),
-        'WZ': ('International Genomics Consortium', 'Testicular Germ Cell Tumors', 'NCH'),
-        'X2': ('University of Washington', 'Sarcoma', 'NCH'),
-        'X3': ('Cleveland Clinic Foundation', 'Testicular Germ Cell Tumors', 'NCH'),
-        'X4': ('Institute for Medical Research', 'Prostate adenocarcinoma', 'NCH'),
-        'X5': ('Institute of Human Virology Nigeria', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'X6': ('University of Iowa', 'Sarcoma', 'NCH'),
-        'X7': ('ABS IUPUI', 'Thymoma', 'NCH'),
-        'X8': ('St. Joseph\'s Hospital Arizona', 'Esophageal carcinoma ', 'NCH'),
-        'X9': ('University of California, Davis', 'Sarcoma', 'NCH'),
-        'XA': ('University of Minnesota', 'Prostate adenocarcinoma', 'NCH'),
-        'XB': ('Albert Einstein Medical Center', 'Esophageal carcinoma ', 'NCH'),
-        'XC': ('Albert Einstein Medical Center', 'Lung squamous cell carcinoma', 'NCH'),
-        'XD': ('Providence Portland Medical Center', 'Pancreatic adenocarcinoma', 'NCH'),
-        'XE': ('University of Southern California', 'Testicular Germ Cell Tumors', 'NCH'),
-        'XF': ('University of Southern California', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'XG': ('BLN UT Southwestern Medical Center at Dallas', 'Pheochromocytoma and Paraganglioma', 'NCH'),
-        'XH': ('BLN Baylor', 'Thymoma', 'NCH'),
-        'XJ': ('University of Kansas', 'Prostate adenocarcinoma', 'NCH'),
-        'XK': ('Mayo Clinic Arizona', 'Prostate adenocarcinoma', 'NCH'),
-        'XM': ('MSKCC', 'Thymoma', 'NCH'),
-        'XN': ('University of Sao Paulo', 'Pancreatic adenocarcinoma', 'NCH'),
-        'XP': ('University of Sao Paulo', 'Esophageal carcinoma ', 'NCH'),
-        'XQ': ('University of Sao Paulo', 'Prostate adenocarcinoma', 'NCH'),
-        'XR': ('University of Sao Paulo', 'Liver hepatocellular carcinoma', 'NCH'),
-        'XS': ('University of Sao Paulo', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'XT': ('Johns Hopkins', 'Mesothelioma', 'NCH'),
-        'XU': ('University Health Network', 'Thymoma', 'NCH'),
-        'XV': ('Capital Biosciences', 'Skin Cutaneous Melanoma', 'NCH'),
-        'XX': ('Spectrum Health', 'Breast invasive carcinoma', 'NCH'),
-        'XY': ('Spectrum Health', 'Testicular Germ Cell Tumors', 'NCH'),
-        'Y3': ('University of New Mexico', 'Acute Myeloid Leukemia', 'NCH'),
-        'Y5': ('University of Arizona', 'Sarcoma', 'NCH'),
-        'Y6': ('University of Arizona', 'Prostate adenocarcinoma', 'NCH'),
-        'Y8': ('Spectrum Health', 'Kidney renal papillary cell carcinoma', 'NCH'),
-        'YA': ('Spectrum Health', 'Liver hepatocellular carcinoma', 'NCH'),
-        'YB': ('Spectrum Health', 'Pancreatic adenocarcinoma', 'NCH'),
-        'YC': ('Spectrum Health', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'YD': ('Spectrum Health', 'Skin Cutaneous Melanoma', 'NCH'),
-        'YF': ('University of Puerto Rico', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'YG': ('University of Puerto Rico', 'Skin Cutaneous Melanoma', 'NCH'),
-        'YH': ('Stanford University', 'Pancreatic adenocarcinoma', 'NCH'),
-        'YJ': ('Stanford University', 'Prostate adenocarcinoma', 'NCH'),
-        'YL': ('PROCURE Biobank', 'Prostate adenocarcinoma', 'NCH'),
-        'YN': ('University of Arizona', 'Skin Cutaneous Melanoma', 'NCH'),
-        'YR': ('Barretos Cancer Hospital', 'Cholangiocarcinoma', 'NCH'),
-        'YS': ('Barretos Cancer Hospital', 'Mesothelioma', 'NCH'),
-        'YT': ('Barretos Cancer Hospital', 'Thymoma', 'NCH'),
-        'YU': ('Barretos Cancer Hospital', 'Testicular Germ Cell Tumors', 'NCH'),
-        'YV': ('MSKCC', 'Uveal Melanoma', 'NCH'),
-        'YW': ('Albert Einstein Medical Center', 'Sarcoma', 'NCH'),
-        'YX': ('Emory University', 'Stomach adenocarcinoma', 'NCH'),
-        'YY': ('Roswell Park', 'Pancreatic adenocarcinoma', 'NCH'),
-        'YZ': ('The Ohio State University', 'Uveal Melanoma', 'NCH'),
-        'Z2': ('IDI-IRCCS', 'Skin Cutaneous Melanoma', 'NCH'),
-        'Z3': ('UCLA', 'Sarcoma', 'NCH'),
-        'Z4': ('Cureline', 'Sarcoma', 'NCH'),
-        'Z5': ('Cureline', 'Pancreatic adenocarcinoma', 'NCH'),
-        'Z6': ('Cureline', 'Esophageal carcinoma ', 'NCH'),
-        'Z7': ('John Wayne Cancer Center', 'Breast invasive carcinoma', 'NCH'),
-        'Z8': ('John Wayne Cancer Center', 'Pancreatic adenocarcinoma', 'NCH'),
-        'ZA': ('Candler', 'Stomach adenocarcinoma', 'NCH'),
-        'ZB': ('Thoraxklinik', 'Thymoma', 'NCH'),
-        'ZC': ('University of Mannheim', 'Thymoma', 'NCH'),
-        'ZD': ('ILSbio', 'Cholangiocarcinoma', 'NCH'),
-        'ZE': ('Spectrum Health', 'Lung squamous cell carcinoma', 'NCH'),
-        'ZF': ('University of Sheffield', 'Bladder Urothelial Carcinoma', 'NCH'),
-        'ZG': ('University Medical Center Hamburg-Eppendorf', 'Prostate adenocarcinoma', 'NCH'),
-        'ZH': ('University of North Carolina', 'Cholangiocarcinoma', 'NCH'),
-        'ZJ': ('NCI HRE Branch', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-        'ZK': ('University of New Mexico', 'Cholangiocarcinoma', 'NCH'),
-        'ZL': ('Valley Hospital', 'Thymoma', 'NCH'),
-        'ZM': ('University of Ulm', 'Testicular Germ Cell Tumors', 'NCH'),
-        'ZN': ('Brigham and Women\'s Hospital Division of Thoracic Surgery', 'Mesothelioma', 'NCH'),
-        'ZP': ('Medical College of Wisconsin', 'Liver hepatocellular carcinoma', 'NCH'),
-        'ZQ': ('Tayside Tissue Bank', 'Stomach adenocarcinoma', 'NCH'),
-        'ZR': ('Tayside Tissue Bank', 'Esophageal carcinoma ', 'NCH'),
-        'ZS': ('Tayside Tissue Bank', 'Liver hepatocellular carcinoma', 'NCH'),
-        'ZT': ('International Genomics Consortium', 'Thymoma', 'NCH'),
-        'ZU': ('Spectrum Health', 'Cholangiocarcinoma', 'NCH'),
-        'ZW': ('University of Alabama', 'Pancreatic adenocarcinoma', 'NCH'),
-        'ZX': ('University of Alabama', 'Cervical squamous cell carcinoma and endocervical adenocarcinoma', 'NCH'),
-    }
-
-    SAMPLE_TYPE = {
-        # 'Code': ('Definition', 'Short Letter Code'),
-        '01': ('Primary solid Tumor', 'TP'),
-        '02': ('Recurrent Solid Tumor', 'TR'),
-        '03': ('Primary Blood Derived Cancer - Peripheral Blood', 'TB'),
-        '04': ('Recurrent Blood Derived Cancer - Bone Marrow', 'TRBM'),
-        '05': ('Additional - New Primary', 'TAP'),
-        '06': ('Metastatic', 'TM'),
-        '07': ('Additional Metastatic', 'TAM'),
-        '08': ('Human Tumor Original Cells', 'THOC'),
-        '09': ('Primary Blood Derived Cancer - Bone Marrow', 'TBM'),
-        '10': ('Blood Derived Normal', 'NB'),
-        '11': ('Solid Tissue Normal', 'NT'),
-        '12': ('Buccal Cell Normal', 'NBC'),
-        '13': ('EBV Immortalized Normal', 'NEBV'),
-        '14': ('Bone Marrow Normal', 'NBM'),
-        '20': ('Control Analyte', 'CELLC'),
-        '40': ('Recurrent Blood Derived Cancer - Peripheral Blood', 'TRB'),
-        '50': ('Cell Lines', 'CELL'),
-        '60': ('Primary Xenograft Tissue', 'XP'),
-        '61': ('Cell Line Derived Xenograft Tissue', 'XCL'),
-    }
