+++ tox.ini
@@ -1,0 +1,105 @@
+[tox]
+envlist =
+  py{27,35,36,37}
+  flake8
+#  lintclient
+
+[testenv]
+deps =
+  coverage
+  mock
+  pytest
+  pytest-cov
+  pytest-girder>=3.0.5
+  pytest-xdist
+  -rrequirements-dev.txt
+install_command = pip install --upgrade --upgrade-strategy eager --find-links https://girder.github.io/large_image_wheels {opts} . {packages}
+whitelist_externals =
+  rm
+  npx
+commands =
+  rm -rf .tox/test/coverage/web_temp
+  girder build --dev
+  pytest --cov {envsitepackagesdir}/digital_slide_archive {posargs}
+#  npx nyc report --temp-dir .tox/test/coverage/web_temp --report-dir .tox/test/coverage --reporter cobertura --reporter text-summary
+
+[testenv:flake8]
+skipsdist = true
+skip_install = true
+deps =
+  flake8
+  flake8-blind-except; python_version >= '3.5'
+  flake8-bugbear
+  flake8-docstrings
+  flake8-quotes
+  pep8-naming
+commands =
+  flake8 {posargs}
+
+[testenv:release]
+passenv =
+  TWINE_USERNAME
+  TWINE_PASSWORD
+  TWINE_REPOSITORY_URL
+deps =
+  twine
+commands =
+  twine check {distdir}/*
+  twine upload --skip-existing {distdir}/*
+
+[flake8]
+max-line-length = 100
+show-source = True
+format = pylint
+max-complexity = 14
+exclude =
+  build
+  node_modules
+  */*egg*/*
+  .git
+  __pycache__
+  .tox
+  */web_client/*
+ignore =
+  # D10* - Missing docstring in *
+  D10
+  D200
+  D205
+  D400
+  D401
+  # E123 - Closing bracket does not match indentation of opening bracket’s line
+  E123
+  # N802 - Function name should be lowercase.
+  N802
+  # N803 - Argument name should be lowercase.
+  N803
+  # N806 - Variable in function should be lowercase.
+  N806
+  # N812 - Lowercase imported as non lowercase.
+  N812
+  # N815 - mixedCase variable in class scope
+  N815
+  # N816 - mixedCase variable in global scope
+  N816
+  # W504 - Line break after binary operator
+  W504
+
+[pytest]
+addopts = --verbose --strict --showlocals --cov-report="term" --cov-report="xml" --cov
+testpaths = tests
+
+[coverage:paths]
+source =
+  digital_slide_archive/
+  .tox/*/lib/python*/site-packages/digital_slide_archive/
+
+[coverage:run]
+data_file = .tox/coverage/.coverage
+branch = True
+omit =
+  tests/*
+  .tox/*
+parallel = True
+
+[coverage:xml]
+output = .tox/coverage/py_coverage.xml
