--- web_client/main.js
@@ -1,1 +1,0 @@
-import './configView';
