--- web_client/templates/digitalSlideArchiveConfig.pug
@@ -1,21 +1,0 @@
-.g-config-breadcrumb-container
-
-form#g-tcga-ingest-form(role="form")
-  p.g-tcga-ingest-description
-    | Ingest can add TCGA files to a collection.
-  .form-group
-    label(for="g-tcga-ingest-datasource") Data Source
-    select#g-tcga-ingest-datasource.form-control
-      option(value='tcga') TCGA
-  .form-group(title="You must ingest TCGA files to a filesystem assetstore.")
-    label(for="g-tcga-ingest-assetstore-id") Assetstore ID
-    input#g-tcga-ingest-assetstore-id.form-control.input-sm(
-      type="text", placeholder="Optional assetstore ID where ingested items will be stored")
-  .form-group
-    label(for="g-tcga-ingest-amount") Number of Items to Ingest
-    input#g-tcga-ingest-amount.form-control.input-sm(
-      type="text", value="10", placeholder="Use the word 'all' to ingest all items, otherwise a subset of items will be ingested")
-  #g-tcga-ingest-error-message
-  button.g-submit-tcga-ingest.btn.btn-success(type="submit")
-    i.icon-link-ext
-    |  Begin Ingest
