--- web_client/templates/layoutFooter.pug
@@ -1,9 +1,0 @@
-.g-footer-links
-  a#g-contact-link(href="mailto:kitware@kitware.com") Contact
-  a(href=apiRoot) Web API
-  a#g-bug-link(href="https://github.com/DigitalSlideArchive/digital_slide_archive/issues/new") Report a bug
-
-.g-footer-info.
-  Powered by
-   #[a#g-about-link(href="http://girder.readthedocs.org/en/latest/user-guide.html") Girder]
-  #[br] &copy; Kitware, Inc.
